+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [9.0.4] - Unreleased
+## [9.1.0] - Unreleased
 - Fix bug with duplicate searchables
+- Add the ability to Template Cohort Identification Configurations (see [Documentation\cohorts\CohortIdentificationConfigurationTemplates.md])
 - Require all deletes to enter a commit message when using the commit system
 - Introduce ability to view Catalogues in a flat view
 - Fix bug where Internal catalogues were still able to be added to a CIC